requirements.txt
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Flask==2.0.1
Flask-Cors==3.0.10
imageio==2.9.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.3
natsort==7.1.1
networkx==2.6.2
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
PyWavelets==1.1.1
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
tifffile==2021.8.8
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
